Updated various files for webpack
