+++ index.html
@@ -30,10 +30,10 @@
         </div>
        <div id="add-new">
         <h1>Add a new book</h1>
-        <form action="">
+        <form>
             <input type="text" id="title" placeholder="title">
             <input type="text" id="author" placeholder="author"> 
-            <div id="formbut"><button id="addbtn" onclick="a.addBooks()"> Add</button></div>
+            <div id="formbut"><button id="addbtn"> Add</button></div>
         </form>
        </div>  
     </section>
@@ -58,7 +58,6 @@
  <footer>
     <p>Copyright</p>
  </footer>
-    
-    <script src="./index.js"></script>
+    <script type="module" src="dist/bundle.js"></script>
 </body>
 </html>
