+++ webpack.config.js
@@ -1,0 +1,11 @@
+const webpack = require('webpack');
+
+import path from 'path';
+
+export default {
+  entry: './index.js',
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(process.cwd(), 'dist'),
+  },
+};
